Integrated J2ME version of the library as of rev 12002
git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@12003 5e8d7ff9-d8ef-0310-90f0-a4852d11357a

+++ src/library/scala/collection/immutable/HashSet.scala
@@ -88,8 +88,8 @@ class HashSet[A] extends Set[A] with mutable.FlatHashTable[A] {
     makeCopyIfUpdated()
     super.elements
   }
-  
-  private def logLimit: Int = Math.sqrt(table.length.toDouble).toInt
+
+  private def logLimit: Int = Math.sqrt(table.length).toInt
 
   private def markUpdated(elem: A, del: Boolean) { 
     val lv = loadFactor
